PhysX 5 update
PhysX 5 update

+++ PEEL/Src/PINT_Common/PINT_CommonPhysX3_ActorManager.cpp
@@ -61,7 +61,7 @@ void ActorManager::Add(PxRigidActor* actor)
 			}
 		}
 	}
-	actor->userData = reinterpret_cast<void*>(Index);
+	actor->userData = reinterpret_cast<void*>(size_t(Index));
 
 	mCachedTimestamp = INVALID_ID;
 }
@@ -88,7 +88,7 @@ bool ActorManager::Remove(PxRigidActor* actor)
 	PxRigidActor* MovedActor = AD[Index].mActor;
 	if(MovedActor!=actor)
 	{
-		ASSERT(udword(MovedActor->userData)==LastIndex);
+		ASSERT(udword(size_t(MovedActor->userData))==LastIndex);
 		MovedActor->userData = reinterpret_cast<void*>(Index);
 	}
 	Actors->ForceSize(LastIndex*ACTOR_DATA_SIZE);
